requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.33.0=pypi_0
aiohappyeyeballs=2.3.4=pypi_0
aiohttp=3.10.0=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.4.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
beartype=0.19.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
bitsandbytes=0.43.3=pypi_0
blas=1.0=mkl
blobfile=3.0.0=pypi_0
bottleneck=1.3.7=py39ha9d4c09_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py39h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
cairo=1.16.0=hb05425b_5
certifi=2025.4.26=py39h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=pypi_0
clip-anytorch=2.6.0=pypi_0
cmake=3.31.0.1=pypi_0
contourpy=1.2.0=py39hdb19cb5_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.5=3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
datasets=2.20.0=pypi_0
dbus=1.13.18=hb2f20db_0
diffusers=0.30.0=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
docker-pycreds=0.4.0=pypi_0
eigen=3.4.0=hdb19cb5_0
einops=0.8.0=pypi_0
ema-pytorch=0.7.3=pypi_0
et_xmlfile=1.1.0=py39h06a4308_0
exceptiongroup=1.2.2=pypi_0
expat=2.6.2=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.13.1=py39h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.51.0=py39h5eee18b_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
ftfy=6.2.0=pypi_0
gdown=5.2.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
grad-cam=1.5.4=pypi_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h11=0.14.0=pypi_0
harfbuzz=4.3.0=hf52aaf7_2
hdf5=1.12.1=h2b7332f_3
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.31.2=pypi_0
icu=73.1=h6a678d5_0
idna=3.7=py39h06a4308_0
imagen-hub=0.2.6=pypi_0
imagen-pytorch=2.1.0=pypi_0
importlib-metadata=8.2.0=pypi_0
importlib_resources=6.4.0=py39h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py39h06a4308_0
joblib=1.4.2=py39h06a4308_0
jpeg=9e=h5eee18b_2
kagglehub=0.3.12=pypi_0
kiwisolver=1.4.4=py39h6a678d5_0
kornia=0.7.3=pypi_0
kornia-rs=0.1.5=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.10.1.7=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.6.82=0
libcurl=8.7.1=h251f7ec_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hfdd30dd_2
lightning-utilities=0.11.6=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lpips=0.1.4=pypi_0
lxml=5.3.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.8.4=py39h06a4308_0
matplotlib-base=3.8.4=py39h1128e8f_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py39h06a4308_0
numexpr=2.8.7=py39h85018f9_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
omegaconf=2.3.0=pypi_0
open-clip-torch=2.30.0=pypi_0
openai=1.37.1=pypi_0
openai-clip=1.0.1=pypi_0
opencv=4.10.0=py39h0a8ef67_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_2
openpyxl=3.1.5=py39h5eee18b_1
openssl=3.0.16=h5eee18b_0
packaging=24.1=py39h06a4308_0
pandas=2.2.2=py39h6a678d5_0
pcre2=10.42=hebb0a14_1
pillow=10.4.0=py39h5eee18b_0
pip=24.0=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
platformdirs=4.3.6=pypi_0
ply=3.11=py39h06a4308_0
protobuf=5.28.3=pypi_0
psutil=6.0.0=pypi_0
pyarrow=17.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pybind11-abi=4=hd3eb1b0_1
pycountry=24.6.1=pypi_0
pycryptodomex=3.21.0=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.15.10=py39h6a678d5_0
pyqt5-sip=12.13.0=py39h5eee18b_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.19=h955ad1f_1
python-dateutil=2.9.0post0=py39h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.4.0=py3.9_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-lightning=2.4.0=pypi_0
pytorch-mutex=1.0=cuda
pytorch-warmup=0.1.1=pypi_0
pytz=2024.1=py39h06a4308_0
pyyaml=6.0.1=py39h5eee18b_0
qt-main=5.15.2=h53bd1ea_10
rapidfuzz=3.12.1=py39h6a678d5_0
readline=8.2=h5eee18b_0
regex=2024.7.24=pypi_0
requests=2.32.3=py39h06a4308_0
safetensors=0.4.3=pypi_0
scikit-learn=1.5.1=py39h1128e8f_0
scipy=1.13.1=py39h5f9d8c6_0
sentence-transformers=5.0.0=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.18.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=py39h06a4308_0
sip=6.7.12=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=py39h2f386ee_0
timm=1.0.14=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.1=pypi_0
tomli=2.0.1=py39h06a4308_0
torchaudio=2.4.0=py39_cu118
torchmetrics=1.4.1=pypi_0
torchtriton=3.0.0=py39
torchvision=0.19.0=py39_cu118
tornado=6.4.1=py39h5eee18b_0
tqdm=4.66.4=pypi_0
transformers=4.51.3=pypi_0
ttach=0.0.3=pypi_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py39h5eee18b_0
urllib3=2.2.2=py39h06a4308_0
wandb=0.18.6=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.43.0=py39h06a4308_0
x264=1!157.20191217=h7b6447c_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zipp=3.17.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
